hydra2: Fixed some compiler warnings

Approved-by: Hui Zhou <hzhou321@anl.gov>

+++ src/pm/hydra2/proxy/proxy_pmi.c
@@ -82,7 +82,7 @@ HYD_status proxy_process_pmi_cb(int fd, HYD_dmx_event_t events, void *userp)
             pmi_stash[count - 1] = 0;
             delim = " ";
         } else if (!strncmp(pmi_stash, "mcmd=", strlen("mcmd="))) {
-            if (count < strlen("endcmd") ||
+            if (count < (int) strlen("endcmd") ||
                 strncmp(&pmi_stash[count - strlen("endcmd")], "endcmd", strlen("endcmd"))) {
                 /* if we reached the end of the buffer we read and did
                  * not yet get a full comamnd, wait till we get at
